requirements.txt
streamlit
requests
python-dotenv
PyPDF2
faiss-cpu
sentence-transformers
